+++ app/assets/stylesheets/components/_btn.scss
@@ -17,3 +17,23 @@
     text-decoration: none;
   }
 }
+
+.link-show {
+  color: $green;
+  padding: 5px;
+  font-size: 24px;
+  transition: 0.5s ease;
+  &:hover {
+    text-decoration: none;
+    color: $blue;
+  }
+}
+
+.tab-nav {
+  color: $green;
+  font-size: 24px;
+}
+
+.ul-link-show {
+  margin: 0 !important;
+}
