+++ app/assets/stylesheets/components/_footer.scss
@@ -1,28 +1,3 @@
-// .footer {
-//   background: white;
-//   display: flex;
-//   align-items: center;
-//   justify-content: space-between;
-//   height: 100px;
-//   padding: 0px 50px;
-//   // color: rgba(0,0,0,0.3);
-// }
-// .footer-links {
-//   display: flex;
-//   align-items: center;
-// }
-// .footer-links a {
-//   color: $blue;
-//   text-decoration: none;
-//   font-size: 24px;
-//   padding: 0px 10px;
-// }
-// .footer-links a:hover {
-//   opacity: 0.8;
-// }
-// .footer .fa-heart {
-//   color: $blue;
-// }
 .full-footer{
   background-color: #0B4F6C;
 }
