+++ app/assets/stylesheets/components/_index.scss
@@ -7,3 +7,4 @@
 @import "footer";
 @import "btn";
 @import "map";
+@import "show_doc"
