+++ app/assets/stylesheets/components/_map.scss
@@ -1,4 +1,4 @@
 #mapid {
-  height: 300px;
-  border-radius: 3px;
+  height: 60vh;
+  border-left: 3px solid $blue;
 }
