+++ app/assets/stylesheets/config/_colors.scss
@@ -5,7 +5,7 @@ $red: #EE5F5B;
 $blue: #0B4F6C;
 $yellow: #FDB631;
 $orange: #E67E22;
-$green: #20BF55;
+$green: #00ab9d;
 $gray: #000000;
 $light-gray: #F4F4F4;
 $white: #FBFBFF;
