+++ app/controllers/document_types_controller.rb
@@ -1,3 +1,2 @@
 class DocumentTypesController < ApplicationController
-
 end
