+++ app/views/houses/index.html.erb
@@ -1,10 +1,6 @@
-<div class="wrapper">
-  <div class="container">
-    <% if current_user %>
-      <%= render 'shared/numbers-houses' %>
-      <%= render 'shared/listing-houses' %>
-      <% else %>
-      <h1> Merci de vous connecter </h1>
-    <% end %>
-  </div>
-</div>
+<% if current_user %>
+  <%= render 'shared/numbers-houses' %>
+  <%= render 'shared/listing-houses' %>
+  <% else %>
+  <h1> Merci de vous connecter </h1>
+<% end %>
