+++ app/views/houses/new.html.erb
@@ -1,11 +1,9 @@
 <!-- new.html.erb -->
-<div class="container">
-  <div class="row">
-    <div class="col-sm-6">
-      <%= render 'form_houses', house: @house %>
-    </div>
-    <div class="col-sm-6 form-right">
-      <img src="https://images.pexels.com/photos/186077/pexels-photo-186077.jpeg?auto=compress&cs=tinysrgb&h=600&w=600" alt="">
-    </div>
+<div class="row">
+  <div class="col-sm-6">
+    <%= render 'form_houses', house: @house %>
+  </div>
+  <div class="col-sm-6 form-right">
+    <img src="https://images.pexels.com/photos/186077/pexels-photo-186077.jpeg?auto=compress&cs=tinysrgb&h=600&w=600" alt="">
   </div>
 </div>
