+++ app/views/pages/home.html.erb
@@ -1,5 +1,3 @@
-
-
 <div class="banner" style="background-image: linear-gradient(-225deg, rgba(0,101,168,0.6) 0%, rgba(0,36,61,0.6) 50%), url(<%= image_url('banner-home-02.jpg') %>);">
   <div class="banner-content">
     <h1>START IMMO</h1>
