+++ app/views/shared/_document-juridiques.html.erb
@@ -1,4 +1,4 @@
-<div class="white-back">
+<div class="white-back" id="sticky-doc">
   <h3 class="background-blue">Document juridiques</h3>
   <ul>
     <% @house.documents.each do |document| %>
