+++ app/views/shared/_footer.html.erb
@@ -1,57 +1,36 @@
-<!-- <div class="footer">
-  <div class="footer-links">
-    <%= link_to "" do %>
-      <i class="fab fa-facebook"></i>
-    <% end%>
-    <%= link_to "" do %>
-      <i class="fab fa-linkedin"></i>
-    <% end%>
-  </div>
-  <div class="footer-copyright">
-    Start Immo <i class="fa fa-heart"></i>
-  </div>
-</div> -->
-
-
-<div class="container full-footer">
-    <div class="row">
-<!--         <div class="col-sm-3">
-            <h2 class="logo"><a href="#"> LOGO </a></h2>
-        </div> -->
-        <div class="col-sm-3 columns-footer">
-            <h5>Get started</h5>
-            <ul>
-                <li><a href="#">Home</a></li>
-                <li><a href="#">Sign up</a></li>
-                <li><a href="#">Downloads</a></li>
-            </ul>
-        </div>
-        <div class="col-sm-3 columns-footer">
-            <h5>About us</h5>
-            <ul>
-                <li><a href="#">Politique de confidentialité</a></li>
-                <li><a href="#">Mentions légales</a></li>
-                <li><a href="#">CGV et CGU</a></li>
-            </ul>
-        </div>
-        <div class="col-sm-3 columns-footer">
-            <h5>A propos</h5>
-            <ul>
-                <li><a href="#">Qui sommes nous</a></li>
-                <li><a href="#">Blog</a></li>
-                <li><a href="#">Forums</a></li>
-            </ul>
-        </div>
-        <div class="col-sm-3 columns-footer">
-            <div class="social-networks">
-                <a href="#" class="twitter"><i class="fab fa-twitter fa-2x"></i>
-</a>
-                <a href="#" class="facebook"><i class="fab fa-facebook-f fa-2x"></i>
-</a>
-                <a href="#" class="google"><i class="fab fa-google-plus-g fa-2x"></i>
-</a>
-            </div>
-            <button type="button" class="btn btn-default">Contactez-nous</button>
-        </div>
+<div class="full-footer">
+  <div class="row reset-boot">
+    <div class="col-sm-3 columns-footer">
+      <h5>Get started</h5>
+      <ul>
+        <li><a href="#">Home</a></li>
+        <li><a href="#">Sign up</a></li>
+        <li><a href="#">Downloads</a></li>
+      </ul>
     </div>
+    <div class="col-sm-3 columns-footer">
+      <h5>About us</h5>
+      <ul>
+        <li><a href="#">Politique de confidentialité</a></li>
+        <li><a href="#">Mentions légales</a></li>
+        <li><a href="#">CGV et CGU</a></li>
+      </ul>
+    </div>
+    <div class="col-sm-3 columns-footer">
+      <h5>A propos</h5>
+      <ul>
+        <li><a href="#">Qui sommes nous</a></li>
+        <li><a href="#">Blog</a></li>
+        <li><a href="#">Forums</a></li>
+      </ul>
+    </div>
+    <div class="col-sm-3 columns-footer">
+      <div class="social-networks">
+        <a href="#" class="twitter"><i class="fab fa-twitter fa-2x"></i></a>
+        <a href="#" class="facebook"><i class="fab fa-facebook-f fa-2x"></i></a>
+        <a href="#" class="google"><i class="fab fa-google-plus-g fa-2x"></i></a>
+      </div>
+      <button type="button" class="btn btn-default">Contactez-nous</button>
+    </div>
+  </div>
 </div>
