+++ app/views/shared/_house-information.html.erb
@@ -1,4 +1,17 @@
 <div class="white-back z-depth-3">
+  <ul class="list-inline text-center ul-link-show">
+    <li>
+      <%= link_to "Informations", house_path, class:"link-show"  %>
+    </li>
+    <li>
+      <span class="tab-nav">|</span>
+    </li>
+    <li>
+      <%= link_to "Historique", house_doc_path(anchor: "link-historique"),  class:"link-show"  %>
+    </li>
+  </ul>
+</div>
+<div class="white-back z-depth-3">
   <div class="title-icon">
     <p>Propriétaire</p>
     <i class="fas fa-male"></i>
@@ -105,4 +118,3 @@
     <div class="wrap-inside">Honnoraires TTC : 10 000€</div>
   </div>
 </div>
-<div id="mapid" class="z-depth-3" data-markers=" <%= @markers.to_json %> "></div>
