+++ app/views/user_mailer/welcome.html.erb
@@ -1,7 +1,5 @@
 <h1>Bienvenue</h1>
 
 <p>
-  Bonjour, chez Start Immo, sécurisez vos transactions immobilères.
+  Bonjour, chez Start Immo : sécurisez vos transactions immobilères.
 </p>
-
-
